French translations, and list of sent_id which are not sentences
sent_id = wikipedia.vislcg.txt:43:1258 is not a sentence
sent_id = wikipedia.vislcg.txt:42:1246 is not a sentence
sent_id = wikipedia.vislcg.txt:34:1001 was missing the end of the ordinal number in the original text.
sent_id = wikipedia.vislcg.txt:31:923 is not a sentence
sent_id = wikipedia.vislcg.txt:6:63 is not a sentence
sent_id = wikipedia.vislcg.txt:5:47 is not a sentence
sent_id = wikipedia.vislcg.txt:4:35 is not a sentence
sent_id = wikipedia.vislcg.txt:3:23 is not a sentence
sent_id = wikipedia.vislcg.txt:2:11 is not a sentence
sent_id = wikipedia.vislcg.txt:1:0 is not a sentence
sent_id = war-bont.vislcg.txt:1:0 is not a sentence, but a temporal relative. I have added the punctuation "..."
sent_id = oab.vislcg.txt:167:4457 is not a sentence but an infinitival verbal syntagma
sent_id = oab.vislcg.txt:160:4269 is not a sentence, but a truncated sentence
sent_id = oab.vislcg.txt:158:4233 is not a sentence
sent_id = oab.vislcg.txt:156:4176 is not a sentence