minor corrections, translations & an ex dropped.
I can not check it on the source paper, but I think that the example # sent_id = grammar.vislcg.txt:49:987 was actually published as ungrammatical (with a star *) and that it is an error to include it here. I have dropped it.
The paper reference is noted as 1993 and it is not her 1993's paper about nominals. The ref. does not appear in [http://talnarchives.atala.org/TALN/TALN-2018/25.pdf the data paper], but I guess it is Stephens, Janig. 1993b. 'Breton', Ball & Fife, The Celtic Languages, 349-409. I don't have access to it.